% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_volume_total}
\alias{calc_volume_total}
\title{Calculate water total volume,using the thermocline depth}
\usage{
calc_volume_total(H, A)
}
\arguments{
\item{H}{depths}

\item{A}{areas}
}
\value{
list of temperatures
}
\description{
Calculate water total volume,using the thermocline depth
}
