% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extract_time_space}
\alias{extract_time_space}
\title{Extract time and space information}
\usage{
extract_time_space(wtemp)
}
\arguments{
\item{wtemp}{matrix; Water temperatures (rows correspond to time, cols to depth)}
}
\value{
list of datetimes and depths
}
\description{
Extracts time (from date column) and space (aka depth) information
}
\examples{
\dontrun{
wtemp <- data.frame(
  date = c("1979-04-02", "1979-04-02"),
  temp_0 = c(0, 0),
  temp_1 = c(1, 1),
  stringsAsFactors = FALSE)
extract_time_space(wtemp)
}
}
